fmt warning c4996 - Buffer.h(62,1): warning C4715: “Hazel::BufferElement::GetComponentCount”: 不是所有的控件路径都返回值 - EditorLayer.cpp(11,91): warning C4244: “参数”: 从“uint32_t”转换到“float”，可能丢失数据

+++ Hazelnut/src/EditorLayer.cpp
@@ -8,7 +8,7 @@ namespace Hazel {
 
 
     EditorLayer::EditorLayer()
-        : Layer("Sandbo2D"), m_CameraController(Application::Get().GetWindow().GetWidth() / Application::Get().GetWindow().GetHeight())
+        : Layer("Sandbo2D"), m_CameraController( (float)Application::Get().GetWindow().GetWidth() / (float)Application::Get().GetWindow().GetHeight())
     {
     }
